Redux (GameState) (Score + Next Level)
I have updated the project to utilize a useGameState Hook that is connected to Redux.

This Hook allows for the updating of the score, gameOver, and speed of the game.

Each level is considered completed when the Score is equal to the maxScore, which is calculated by adding up all the goals in the current level to the previous score.

The speed state in Redux is updated by incrementing it by 0.5.

This update also generates a new map, which can be seen in App.tsx.