Sound Control Component
Write a component called SoundControl that allows users to control whether they want to hear the sound or not. If they don't want to hear the sound, they can click a button to update the Redux Store with the "mute" state.

The Player component, which receives a song (player.tsx), and the App component, which includes the <MusicLoop/> component, have been updated to retrieve the "mute" state from the Redux Store.