Fix: What Happen if we can't decrypt ?
Even if it shouldn't be the case they could be have issue for a reason or a other that someone put in the DB a incorrect "score" encrypted.

Since we decrypt all scores if it fail it stop by reading other scores.

So we put a "try/catch" about crypto.ts that will return  the data or false.

And we check that the name & score match a Regex and that score is a Number.